Added a few more syscalls. I'll be porting some sort of standard library next

+++ src/kernel/elf.c
@@ -21,7 +21,7 @@ char elf_ident[4] = {0x7f, 'E', 'L', 'F'};
 
 // TYSM OSDEV wiki
 static inline elf_sect_header_t *elf_sheader(elf_header_t *hdr) {
-  return (elf_sect_header_t *)((uint32_t)(uint64_t)hdr + hdr->sect_head_off);
+  return (elf_sect_header_t *)((uintptr_t)(uint64_t)hdr + hdr->sect_head_off);
 }
 
 uint8_t elf_run_binary(char *name, char *path, proc_t *proc, size_t time_slice,
